Make a basic Wordgroup
Constructor to make a new WordGroup, putting just a single word into a
linked list. Calculate the lowercase sorted value of the given "word"
and store that. That is the "key" that should be the same for all words
allowed in the group.
Add a method to return the key for a group.
Add tests that create WordGroup(s) with various combinations of
characters in the initial word, and check that the WordKey is sorted as
expected.